% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lapse_study_data}
\alias{lapse_study_data}
\alias{training}
\alias{testing}
\alias{validation}
\title{Lapse study data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 40138 rows and 17 columns.}
\usage{
training

testing

validation
}
\description{
Dataset derived from SOA lapse study data. For details see `data-raw/data-prep.R`.
}
\keyword{datasets}
