% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_keras_data}
\alias{make_keras_data}
\title{Preprocess Data for Keras Lapse Model}
\usage{
make_keras_data(data, recipe)
}
\arguments{
\item{data}{Dataset with categorical variables as factors.}

\item{recipe}{Trained recipe object.}
}
\description{
Preprocess Data for Keras Lapse Model
}
