% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_xgb_data}
\alias{make_xgb_data}
\title{Preprocess data for XGB model}
\usage{
make_xgb_data(data, recipe)
}
\arguments{
\item{data}{Dataset with categorical variables as factors.}

\item{recipe}{Trained recipe object.}
}
\description{
Preprocess data for XGB model
}
